+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Version 1.5.0:
 
 * Modified `grizzled.string.WordWrapper` to allow specification of
   characters that should be ignored when calculating wrapping.
+* Removed deprecated methods from `grizzled.config.Configuration`.
+* Added `+` and `-` methods to `grizzled.config.Configuration`, to allow
+  addition to and removal from a `Configuration` in an immutable way.
 * Updated ScalaTest to 2.2.6 and scala-async to 0.9.5.
 
 Version 1.4.0: